React Native Application
This a simple application that uses firebase authntication to login and logout the user. It checks if the email is new then it creates the account with the entered email and password and if the email is already registered then it validates it that the entered password is correct or not.